Refactor and cleanup code a little.
