--- polar-analyzer/pkg/polar_analyzer.d.ts
@@ -1,23 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-/**
-*/
-export class Polar {
-  free(): void;
-/**
-*/
-  constructor();
-/**
-* @param {string} src
-* @param {string} filename
-* @returns {any}
-*/
-  load(src: string, filename: string): any;
-/**
-*/
-  clearRules(): void;
-/**
-* @returns {any}
-*/
-  getSummary(): any;
-}
