--- polar-analyzer/pkg/polar_analyzer_bg.wasm.d.ts
@@ -1,11 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export function __wbg_polar_free(a: number): void;
-export function polar_wasm_new(): number;
-export function polar_load(a: number, b: number, c: number, d: number, e: number): number;
-export function polar_clearRules(a: number): void;
-export function polar_getSummary(a: number): number;
-export function __wbindgen_malloc(a: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number): number;
-export function __wbindgen_free(a: number, b: number): void;
