+++ polar-analyzer/src/inspect/mod.rs
@@ -1,0 +1,6 @@
+//! Inspect Polar code to extract information about
+//! rules, terms, etc.
+
+mod rules;
+
+pub use rules::{get_rule_info, RuleInfo};
